feature/repeat : 집안일 목록 조회 api 변경에 따른 코드 변경 완료
