+++ app/src/main/java/com/depromeet/housekeeper/data/dataSource/RemoteDataSource.kt
@@ -1,6 +1,5 @@
 package com.depromeet.housekeeper.data.dataSource
 
-import com.depromeet.housekeeper.model.HouseWork
 import com.depromeet.housekeeper.model.request.*
 import com.depromeet.housekeeper.model.response.*
 import kotlinx.coroutines.flow.Flow
@@ -22,7 +21,7 @@ interface RemoteDataSource {
     suspend fun getDateHouseWorkList(
         fromDate: String,
         toDate: String
-    ): Flow<Map<String, HouseWorks>>
+    ): Flow<ApiResult<Map<String, HouseWorks>>>
 
     suspend fun getPeriodHouseWorkListOfMember(
         teamMemberId: Int,
